interpolated wrt time

+++ interpolation/include/interpolation/pose.h
@@ -1,6 +1,6 @@
 struct Pose
 {
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 };
